+++ enclone_tail/src/lib.rs
@@ -18,6 +18,7 @@ pub mod pack_circles;
 pub mod phylip;
 pub mod plot;
 pub mod plot_points;
+pub mod plot_utils;
 pub mod polygon;
 pub mod print_stats;
 pub mod requirements;
